+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 2.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Don't show header viewlet if settings aren't set. Now default settings are set
+  at install-time.
+  [cekk]
 
 
 2.0.0 (2018-03-15)